requirements.txt
discord.py
asyncpg
jishaku
humanfriendly
wavelink
git+https://github.com/Rapptz/discord-ext-menus
beautifulsoup4
pynacl
tortoise-orm
python-dotenv